backend/requirements.txt
fastapi
uvicorn
torchvision
Pillow
pydantic
python-multipart
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
ffmpeg==1.4
filelock==3.18.0
fsspec==2025.5.1
idna==3.10
Jinja2==3.1.6
llvmlite==0.44.0
MarkupSafe==3.0.2
more-itertools==10.7.0
mpmath==1.3.0
networkx==3.4.2
numba==0.61.2
numpy==2.2.6
openai-whisper @ git+https://github.com/openai/whisper.git@dd985ac4b90cafeef8712f2998d62c59c3e62d22
pydub==0.25.1
regex==2024.11.6
requests==2.32.4
semantic-version==2.10.0
setuptools-rust==1.11.1
sympy==1.14.0
tiktoken==0.9.0
torch==2.7.1
tqdm==4.67.1
typing_extensions==4.14.0
urllib3==2.4.0
matplotlib
pandas
pymongo 
transformers
peft 
sentence_transformers
aiofiles
seaborn 
librosa 
pathlib
scikit-learn 
huggingface_hub
